+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## 💡 Enhancements 💡
 
-- `tailsampling` processor: Add new policy `percentage` (#3876)
+- `tailsampling` processor: Add new policy `probabilistic` (#3876)
 
 ## v0.31.0